requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
feedparser==5.2.1
idna==2.6
lxml==4.0.0
mysql-connector==2.1.4
oauthlib==2.0.4
pathlib==1.0.1
py-trello==0.9.0
pypandoc==1.4
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2017.2
Pyzotero==1.2.13
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
Unidecode==0.4.21
urllib3==1.22
